feat(ball): Draw a red ball (rect) on the screen. (#13)
### TL;DR

This PR adds a `Ball` struct to the game and implements the functionality to draw it.

### What changed?

`Ball` struct has been added, and it consists two fields: `radius` and `to_rect` method for converting the ball to a rectangle that can be rendered. `Ball` creation is added to the main loop. Also, functionality to draw the ball on the game window has been added in the main application loop.

### Why make this change?

To enhance gameplay by adding a ball that will bounce around the window and off the paddle.

NOTE: Right now, ball is static.

---